[Refactor] support assets for `GenericMessage` (#952)
